// src/store/types.ts
export interface ITeam {
  id: number;
  imgUrl: string;
  name: string;
  role: string;
}

export interface ISneakers {
  color?: string;
  compound?: string;
  country?: string;
  description?: string;
  gender?: string;
  id?: number;
  imgUrl?: string;
  inStock?: number;
  oldPrice?: number;
  price?: number;
  sizes?: number[];
  stars?: number;
  title?: string;
  vendorСode?: string;
  selectedSize: number;
}